move lib/api to clientApi, comment out likely unused client side api fns
